Add files via upload
There is a daily as well as on-demand need to extract weather details for a list of cities, for the next 7 days so that it could be shared with employees who are on the business trips during this period. The list of cities with dates is provided in a comma separated files. as the weather details for a given city, for a given date or period, is required different automation purposes, we need to create a specialized reusable automation process.